requirements.txt
redis
pymongo
pytest
ultralytics
Pillow
requests
streamlit
python-dotenv